+++ src/core/light/DirectionalLight2.ts
@@ -50,11 +50,14 @@ export class DirectionalLight2 extends DirectionalLight implements ILight<Direct
     @onChange3('setDirty')
     declare castShadow: boolean
 
-    @uiVector('Shadow Map Size')
+    @uiVector('Shadow Map Size', undefined, undefined, (that: DirectionalLight2)=>({onChange: ()=>that._mapSizeChanged()}))
     @bindToValue({obj: 'shadow', key: 'mapSize', onChange: DirectionalLight2.prototype._mapSizeChanged, onChangeParams: false})
         shadowMapSize: Vector2
 
-    protected _mapSizeChanged() {
+    /**
+     * @internal
+     */
+    ['_mapSizeChanged']() {
         this.shadow.map?.dispose()
         this.shadow.mapPass?.dispose()
         this.shadow.map = null as any