+++ src/core/light/SpotLight2.ts
@@ -62,11 +62,14 @@ export class SpotLight2 extends SpotLight implements ILight<SpotLightShadow> {
     @onChange3('setDirty')
     declare castShadow: boolean
 
-    @uiVector('Shadow Map Size')
+    @uiVector('Shadow Map Size', undefined, undefined, (that: SpotLight2)=>({onChange: ()=>that._mapSizeChanged()}))
     @bindToValue({obj: 'shadow', key: 'mapSize', onChange: SpotLight2.prototype._mapSizeChanged, onChangeParams: false})
         shadowMapSize: Vector2
 
-    protected _mapSizeChanged() {
+    /**
+     * @internal
+     */
+    ['_mapSizeChanged']() {
         this.shadow.map?.dispose()
         this.shadow.mapPass?.dispose()
         this.shadow.map = null as any